requirements.txt
cryptography==38.0.1
fastapi==0.85.1
python-dotenv==0.19.2
uvicorn==0.19.0
